Initial draft of the migrateFrom.
The comparision is done on CPU flags (only). You provide the node name that you wish to migrate from, it then retreives
the nova-compute nodes from OpenStack, compares the flags used on the source to the tested nodes. If the flags are
identical its listed as a match, if the node has all the source flags and more, its a covered match. If flags are missing,
then things become complicated. Currently there is no logic for handling missing flags, this includes comparing different
manufacturer flags, cf.  AMD and Intel.

The output is three lines;
perf = nodenames
covered = nodenames
unknown = nodenames